third_party/WebKit/LayoutTests/inspector/styles/keyframes-rules-expected.txt

Tests that source data is extracted correctly from stylesheets with @-webkit-keyframes rules.

Content

Running: testInit

Running: testDumpStyles
[expanded] 
element.style  { ()
background-color: white;

======== Matched CSS Rules ========
[expanded] 
*  { (keyframes-rules.html:9)
/-- overloaded --/ background-color: papayawhip;

[expanded] 
div  { (user agent stylesheet)
display: block;
